// iot-service/iot-angular/src/sdk/core/models/document-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface DocumentResponseDto {
  createdAt: string;
  documentType: string;
  fileExtension: string;
  filePath: string;
  fileSize: number;
  fromModule: string;
  fromModuleId: string;
  idDocument: string;
  idOwner: string;
  metadata?: {
};
  mimeType: string;
  originalFilename: string;
  status: string;
  storedFilename: string;
}