// iot-service/iot-angular/src/sdk/core/models/layer-response-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface LayerResponseDto {
  bbox?: {
};
  categoryCode?: string;
  configJson: {
};
  createdAt: string;
  displayOrder: number;
  featureCount: number;
  geometryType?: string;
  idLayer: string;
  idOwner?: string;
  idProject?: string;
  isCore: boolean;
  isLocked: boolean;
  isVisibleDefault: boolean;
  layerCode?: string;
  layerDescription?: string;
  layerName: string;
  layerType: 'core' | 'operational' | 'custom';
  sourceRef?: string;
  sourceTable?: string;
  sourceType: 'system' | 'geojson' | 'shp' | 'kml' | 'csv' | 'api';
  srid: number;
  styleJson: {
};
  updatedAt: string;
}